"cspp2_m10_a1_1: 0/5 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ cspp2-assignments/Wallet/Wallet.java
@@ -1,6 +1,6 @@
 public class Wallet {
 	String name;
-	float amount = 1000;
+	int amount = 1000;
 	// User[] user;
 	Wallet() {
 		this.amount = 1000;
@@ -11,20 +11,20 @@ public class Wallet {
 	public String getName() {
 		return this.name;
 	}
-	public float getAmount() {
+	public int getAmount() {
 		return this.amount;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public void setAmount(float amount) {
+	public void setAmount(int amount) {
 		this.amount = amount;
 	}
 	public String toString() {
 		return this.name+" "+this.amount;
 	}
 
-	// public void credit(String name, float amount) {
+	// public void credit(String name, int amount) {
 
 	// }
 	// public void debit(String name, int amount) {
